feat: added action term for a keyword, started restructuring into actions_and_functions module
