+++ src/ast/cdl/actions_and_functions/action_term.rs
@@ -1,0 +1,7 @@
+use crate::ast::Location;
+use ast_term_derive::ASTTerm;
+
+#[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
+pub struct ActionTerm {
+    location: Location,
+}
