--- src/parser/single_module/cdl/action/grammar.lalrpop
@@ -1,8 +1,0 @@
-ActionDeclaration: ActionDeclarationTerm = {
-  "action" <name:Name> <open_bracket:OpenRoundBracket> <params:Params> <close_bracket:CloseRoundBracket> ";" => {
-    ActionDeclarationTerm::new(Box::new(name), Box::new(open_bracket), params, Box::new(close_bracket), None)
-  },
-  "action" <name:Name> <open_bracket:OpenRoundBracket> <params:Params> <close_bracket:CloseRoundBracket> "returns" <returns:Returns> ";" => {
-    ActionDeclarationTerm::new(Box::new(name), Box::new(open_bracket), params, Box::new(close_bracket), Some(Box::new(returns)))
-  }
-};
